requirements.txt
notion-client>=2.2.0
openai>=1.14.3
google-api-python-client
tenacity>=8.0.0
python-dateutil>=2.9.0 
python-dotenv>=1.0.1
# Vendored library dependencies:
aiofiles>=23.2.1
aiohttp>=3.9.0
duckduckgo-search==7.5.0
firecrawl-py>=2.4.0 # Keep firecrawl itself
kitchenai-whisk
playwright==1.50.0
prompt-toolkit>=3.0.0
rich>=13.7.0
tiktoken>=0.5.0
typer>=0.9.0
# Add any transitive dependencies if they weren't installed before
streamlit
pyyaml
bcrypt
redis
validators
beautifulsoup4
httpx>=0.27.0 # For async HTTP requests

# For RAG (Retrieval Augmented Generation)
sentence-transformers
faiss-cpu

# Testing
pytest>=7.4.0

# Document processing
PyMuPDF # For PDF processing (fitz)
python-docx # For DOCX processing
pandas>=1.5.0,<3.0.0
